Modifie Product Repo with help function
