+++ src/main/java/model/Order.java
@@ -1,5 +1,6 @@
 package model;
 
+import repo.OrderRepo;
 import repo.ProductRepo;
 
 import java.util.Objects;
@@ -7,6 +8,19 @@ import java.util.Objects;
 public class Order {
     private int orderNumber;
     private ProductRepo allProducts;
+
+    public Order(ProductRepo allProducts, int orderNumber){
+        this.allProducts = allProducts;
+        this.orderNumber = orderNumber;
+    }
+
+    public ProductRepo getAllProducts() {
+        return allProducts;
+    }
+
+    public int getOrderNumber(){
+        return this.orderNumber;
+    }
 
     @Override
     public boolean equals(Object o) {
